TAP-103: Styling rework to keep everything as dry as possible
