+++ packages/article/styles/article-header-style.js
@@ -1,6 +1,6 @@
 import { StyleSheet, Platform } from "react-native";
 import globalStyle from "./article-global-style";
-import { COLOUR_HEADLINE, COLOUR_TEXT } from './const/article-const';
+import { COLOUR_HEADLINE, COLOUR_TEXT } from "./const";
 
 const standFirstFontFamily = "TimesModern-Regular";
 
@@ -10,14 +10,14 @@ const webStyles = StyleSheet.create({
   articleHeadLineText: {
     fontSize: 30,
     lineHeight: 32,
-    color: COLOUR_TEXT,
+    color: COLOUR_HEADLINE,
     marginBottom: 7
   },
   standFirst: {
     fontSize: 21,
     lineHeight: 26,
     fontFamily: standFirstFontFamily,
-    color: COLOUR_HEADLINE,
+    color: COLOUR_TEXT,
     paddingBottom: 9
   }
 });
@@ -26,14 +26,14 @@ export const nativeStyles = StyleSheet.create({
   articleHeadLineText: {
     fontSize: Platform.OS === "android" ? 28 : 30,
     lineHeight: Platform.OS === "android" ? 37 : 32,
-    color: COLOUR_TEXT,
+    color: COLOUR_HEADLINE,
     fontWeight: Platform.OS === "android" ? "400" : "700",
     marginBottom: 7
   },
   standFirst: {
     fontSize: 21,
     lineHeight: Platform.OS === "android" ? 29 : 25,
-    color: COLOUR_HEADLINE,
+    color: COLOUR_TEXT,
     fontFamily: standFirstFontFamily,
     paddingBottom: 9
   }
