--- packages/article/styles/article-style-extension.android.js
@@ -1,11 +1,0 @@
-import { StyleSheet } from "react-native";
-import { BODY_FONT_FAMILY } from './const/article-const';
-
-const styles = StyleSheet.create({
-    articleTextElement: { 
-        fontFamily: BODY_FONT_FAMILY,
-        fontSize: 16,
-    },
-});
-
-export default styles;
