--- packages/article/styles/const/article-const.android.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const BODY_FONT_FAMILY = "TimesDigitalW04-Regular";
