+++ packages/article/styles/const/index.js
@@ -1,3 +1,4 @@
 export const FONT_FAMILY_BODY = "TimesDigitalW04";
 export const COLOUR_TEXT = "#333333";
 export const COLOUR_HEADLINE = "#000000";
+export const COLOUR_BORDER = "#D0CECE";
